Gencates/requirements.txt
#pythonpy
djangorestframework==3.14.0
requests==2.30.0
django-cors-headers==3.14.0

aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
certifi==2024.6.2
frozenlist==1.4.1
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.7.4
pydantic-core==2.18.4
yarl==1.9.4
